Move hooks outside components
